Replacement of filler values, etc.
Entry of complexity and entropy values.  Note that in the sentence
"Finally, SSX magnetic fluctuations have complexity values of...", I
replaced "y4" and "y5" with the complexities of both Bdot points. Let me
know if one was supposed to be the complexity of a B data point.
Potentially confusing mention of maximal complexity difference in
choosing L_wind and tau_wind has been deleted. Figure 1 caption edited
to be more explicit about markers. Rosso citation added for first
mention of stochastic fBm and chaotic maps. "we found to be" added to
sentence ending "to avoid artificial numerical effects associated with
small L to  ratios" in order to clarify the lack of need for citation.
Claim that "this is the first time that developed MHD turbulence in an
astrophysical plasma has been identified based on a complexity-entropy
analysis" corrected to refer to a "complexity-entropy plane analysis".
Various grammatical corrections were also made.